+++ panda/board/safety/safety_chrysler.h
@@ -34,6 +34,7 @@ typedef struct {
   const int ESP_8;
   const int ECM_5;
   const int DAS_3;
+  const int DAS_5;
   const int DAS_6;
   const int LKAS_COMMAND;
   const int LKAS_HEARTBIT;
@@ -46,7 +47,8 @@ const ChryslerAddrs CHRYSLER_ADDRS = {
   .ESP_1            = 0x140,  // Brake pedal and vehicle speed
   .ESP_8            = 0x11C,  // Brake pedal and vehicle speed
   .ECM_5            = 0x22F,  // Throttle position sensor
-  .DAS_3            = 0x1F4,  // ACC engagement states from DASM
+  .DAS_3            = 0x1F4,  // ACC
+  .DAS_5            = 0x271,  // ACC for hybrids
   .DAS_6            = 0x2A6,  // LKAS HUD and auto headlight control from DASM
   .LKAS_COMMAND     = 0x292,  // LKAS controls from DASM
   .LKAS_HEARTBIT    = 0x2D9,  // LKAS HEARTBIT from DASM
@@ -81,6 +83,8 @@ const CanMsg CHRYSLER_TX_MSGS[] = {
   {CHRYSLER_ADDRS.CRUISE_BUTTONS, 0, 3},
   {CHRYSLER_ADDRS.LKAS_COMMAND, 0, 6},
   {CHRYSLER_ADDRS.DAS_6, 0, 8},
+  {CHRYSLER_ADDRS.DAS_3, 0, 8},
+  {CHRYSLER_ADDRS.DAS_5, 0, 8},
   {CHRYSLER_ADDRS.LKAS_HEARTBIT, 0, 5},
 };
 
@@ -190,6 +194,7 @@ static void chrysler_rx_hook(const CANPacket_t *to_push) {
   const int das_3_bus = (chrysler_platform == CHRYSLER_PACIFICA) ? 0 : 2;
   if ((bus == das_3_bus) && (addr == chrysler_addrs->DAS_3)) {
     bool cruise_engaged = GET_BIT(to_push, 21U) || ((alternative_experience & ALT_EXP_AOLC_ENABLED) && GET_BIT(to_push, 20U));
+    cruise_engaged = true; // FIXME: Check if experimental is enabled && !GET_BIT(to_push, 21U)
     pcm_cruise_check(cruise_engaged);
   }