+++ selfdrive/car/chrysler/values.py
@@ -34,7 +34,9 @@ class CAR(Platforms):
   # Chrysler
   CHRYSLER_PACIFICA_2017_HYBRID = ChryslerPlatformConfig(
     [ChryslerCarDocs("Chrysler Pacifica Hybrid 2017")],
-    ChryslerCarSpecs(mass=2242., wheelbase=3.089, steerRatio=16.2),
+    ChryslerCarSpecs(mass=2242., wheelbase=3.089, steerRatio=16.2,
+                     gearRatios=[4.70, 2.84, 1.91, 1.38, 1.00, 0.81, 0.70, 0.58],
+                     axleRatio=3.25),
   )
   CHRYSLER_PACIFICA_2018_HYBRID = ChryslerPlatformConfig(
     [ChryslerCarDocs("Chrysler Pacifica Hybrid 2018")],
@@ -65,7 +67,9 @@ class CAR(Platforms):
   # Jeep
   JEEP_GRAND_CHEROKEE = ChryslerPlatformConfig(  # includes 2017 Trailhawk
     [ChryslerCarDocs("Jeep Grand Cherokee 2016-18", video_link="https://www.youtube.com/watch?v=eLR9o2JkuRk")],
-    ChryslerCarSpecs(mass=1778., wheelbase=2.71, steerRatio=16.7),
+    ChryslerCarSpecs(mass=1778., wheelbase=2.71, steerRatio=16.7,
+                     gearRatios=[4.71, 3.14, 2.10, 1.67, 1.29, 1.00, 0.84, 0.67],
+                     axleRatio=3.45)
   )
 
   JEEP_GRAND_CHEROKEE_2019 = ChryslerPlatformConfig(  # includes 2020 Trailhawk
@@ -106,12 +110,19 @@ class CarControllerParams:
       self.STEER_DELTA_DOWN = 3
       self.STEER_MAX = 261  # higher than this faults the EPS
 
+    self.ACC_CONTROL_STEP = 2  # 50Hz
+
+    self.ACCEL_MIN = -3.5
+    self.ACCEL_MAX = 2.0
+    self.INACTIVE_ACCEL = 4.0
+
 
 STEER_THRESHOLD = 120
 
 RAM_DT = {CAR.RAM_1500_5TH_GEN, }
 RAM_HD = {CAR.RAM_HD_5TH_GEN, }
 RAM_CARS = RAM_DT | RAM_HD
+HYBRID_CARS = {CAR.CHRYSLER_PACIFICA_2017_HYBRID,CAR.CHRYSLER_PACIFICA_2018_HYBRID, CAR.CHRYSLER_PACIFICA_2019_HYBRID}
 
 DRIVE_PERSONALITY = [
   [0, 0, 1, 2],