+++ selfdrive/ui/qt/util.cc
@@ -247,11 +247,11 @@ QPixmap bootstrapPixmap(const QString &id) {
 bool hasLongitudinalControl(const cereal::CarParams::Reader &car_params) {
   // Using the experimental longitudinal toggle, returns whether longitudinal control
   // will be active without needing a restart of openpilot
-  const bool expLongAvailable = false; // CP.getExperimentalLongitudinalAvailable()
-  const bool op_long = car_params.getOpenpilotLongitudinalControl() && !expLongAvailable;
-  const bool exp_long_enabled = expLongAvailable && Params().getBool("ExperimentalLongitudinalEnabled");
+  //  return car_params.getExperimentalLongitudinalAvailable()
+  //             ? Params().getBool("ExperimentalLongitudinalEnabled")
+  //             : car_params.getOpenpilotLongitudinalControl();
 
-  return op_long || exp_long_enabled;
+  return true;
 }
 
 // ParamWatcher